requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Django==2.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-js-asset==1.2.2
django-mptt==0.9.1
django-npm==1.0.0
kombu==4.5.0
markdown2==2.3.7
msgpack==0.6.1
pytz==2018.9
PyYAML==5.1
redis==3.2.1
sqlparse==0.3.0
vine==1.3.0
